Refactor TextView url handling to Extension
