+++ app/src/main/java/com/concordium/wallet/uicore/Extensions.kt
@@ -1,8 +1,15 @@
 package com.concordium.wallet.uicore
 
 import android.text.Editable
+import android.text.SpannableStringBuilder
+import android.text.Spanned
 import android.text.TextWatcher
+import android.text.method.LinkMovementMethod
+import android.text.style.ClickableSpan
+import android.text.style.URLSpan
+import android.view.View
 import android.widget.EditText
+import android.widget.TextView
 
 fun EditText.afterTextChanged(afterTextChanged: (String) -> Unit) {
     this.addTextChangedListener(object : TextWatcher {
@@ -16,4 +23,28 @@ fun EditText.afterTextChanged(afterTextChanged: (String) -> Unit) {
             afterTextChanged.invoke(editable.toString())
         }
     })
+}
+
+fun TextView.handleUrlClicks(onClicked: ((String) -> Unit)? = null) {
+    //create span builder and replaces current text with it
+    text = SpannableStringBuilder.valueOf(text).apply {
+        //search for all URL spans and replace all spans with our own clickable spans
+        getSpans(0, length, URLSpan::class.java).forEach {
+            //add new clickable span at the same position
+            setSpan(
+                object : ClickableSpan() {
+                    override fun onClick(widget: View) {
+                        onClicked?.invoke(it.url)
+                    }
+                },
+                getSpanStart(it),
+                getSpanEnd(it),
+                Spanned.SPAN_INCLUSIVE_EXCLUSIVE
+            )
+            //remove old URLSpan
+            removeSpan(it)
+        }
+    }
+    //make sure movement method is set
+    movementMethod = LinkMovementMethod.getInstance()
 }
